simplify components in favor of styled-components
