+++ package.json
@@ -9,11 +9,13 @@
     "@vx/axis": "^0.0.196",
     "@vx/clip-path": "^0.0.196",
     "@vx/curve": "^0.0.196",
+    "@vx/event": "^0.0.196",
     "@vx/grid": "^0.0.196",
     "@vx/group": "^0.0.196",
     "@vx/responsive": "^0.0.196",
     "@vx/scale": "^0.0.196",
     "@vx/shape": "^0.0.196",
+    "@vx/tooltip": "^0.0.196",
     "@vx/voronoi": "^0.0.196",
     "bluebird": "^3.7.2",
     "csv-parse": "^4.8.8",
